[MODEL-24001] -- fix mcp xaa client module naming repo
